change: relocate files previously moved temporarily under Rendering.V2
