+++ src/Beutl.Engine/Graphics/FilterEffects/EffectTarget.cs
@@ -1,7 +1,6 @@
 ﻿using System.ComponentModel;
 using Beutl.Collections.Pooled;
 using Beutl.Graphics.Rendering;
-using Beutl.Graphics.Rendering.V2;
 using Beutl.Media.Source;
 using SkiaSharp;
