+++ src/Beutl.Engine/Graphics/ICanvas.cs
@@ -1,5 +1,5 @@
 ﻿using Beutl.Graphics.Effects;
-using Beutl.Graphics.Rendering.V2;
+using Beutl.Graphics.Rendering;
 using Beutl.Media;
 using Beutl.Media.Pixel;
 using Beutl.Media.Source;
