+++ src/Beutl.Engine/Graphics/IImmediateCanvasFactory.cs
@@ -1,4 +1,4 @@
-﻿using Beutl.Graphics.Rendering.V2.Cache;
+﻿using Beutl.Graphics.Rendering.Cache;
 using SkiaSharp;
 
 namespace Beutl.Graphics;
