+++ src/Beutl.Engine/Graphics/Rendering/DrawBackdropRenderNode.cs
@@ -1,4 +1,4 @@
-﻿namespace Beutl.Graphics.Rendering.V2;
+﻿namespace Beutl.Graphics.Rendering;
 
 public class DrawBackdropRenderNode(IBackdrop backdrop, Rect bounds) : RenderNode()
 {
