+++ src/Beutl.Engine/Graphics/Rendering/OpacityMaskRenderNode.cs
@@ -1,6 +1,6 @@
 ﻿using Beutl.Media;
 
-namespace Beutl.Graphics.Rendering.V2;
+namespace Beutl.Graphics.Rendering;
 
 public sealed class OpacityMaskRenderNode(IBrush mask, Rect maskBounds, bool invert) : ContainerRenderNode
 {
