+++ src/Beutl.Engine/Graphics/Rendering/OpacityRenderNode.cs
@@ -1,4 +1,4 @@
-﻿namespace Beutl.Graphics.Rendering.V2;
+﻿namespace Beutl.Graphics.Rendering;
 
 public sealed class OpacityRenderNode(float opacity) : ContainerRenderNode
 {
