+++ src/Beutl.Engine/Graphics/Rendering/RectClipRenderNode.cs
@@ -1,4 +1,4 @@
-﻿namespace Beutl.Graphics.Rendering.V2;
+﻿namespace Beutl.Graphics.Rendering;
 
 public sealed class RectClipRenderNode(Rect clip, ClipOperation operation) : ContainerRenderNode
 {
