+++ src/Beutl.Engine/Graphics/Rendering/RectangleRenderNode.cs
@@ -1,6 +1,6 @@
 ﻿using Beutl.Media;
 
-namespace Beutl.Graphics.Rendering.V2;
+namespace Beutl.Graphics.Rendering;
 
 public sealed class RectangleRenderNode(Rect rect, IBrush? fill, IPen? pen)
     : BrushRenderNode(fill, pen)
