+++ src/Beutl.Engine/Graphics/Rendering/RenderNode.cs
@@ -1,4 +1,4 @@
-﻿namespace Beutl.Graphics.Rendering.V2;
+﻿namespace Beutl.Graphics.Rendering;
 
 public abstract class RenderNode : INode
 {
