+++ src/Beutl.Engine/Graphics/Rendering/Renderer.cs
@@ -1,5 +1,5 @@
 ﻿using Beutl.Animation;
-using Beutl.Graphics.Rendering.V2.Cache;
+using Beutl.Graphics.Rendering.Cache;
 using Beutl.Media;
 using Beutl.Media.Pixel;
 using SkiaSharp;
