+++ src/Beutl.Engine/Graphics/Rendering/TextRenderNode.cs
@@ -2,7 +2,7 @@
 using Beutl.Media.TextFormatting;
 using SkiaSharp;
 
-namespace Beutl.Graphics.Rendering.V2;
+namespace Beutl.Graphics.Rendering;
 
 public sealed class TextRenderNode(FormattedText text, IBrush? fill, IPen? pen)
     : BrushRenderNode(fill, pen)
