+++ src/Beutl.Engine/Graphics/Shapes/Shape.cs
@@ -1,7 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using Beutl.Animation;
 using Beutl.Graphics.Rendering;
-using Beutl.Graphics.Rendering.V2;
 using Beutl.Language;
 using Beutl.Media;
