+++ src/Beutl/ViewModels/PlayerViewModel.cs
@@ -7,8 +7,7 @@ using Beutl.Audio.Platforms.XAudio2;
 using Beutl.Configuration;
 using Beutl.Graphics;
 using Beutl.Graphics.Rendering;
-using Beutl.Graphics.Rendering.V2;
-using Beutl.Graphics.Rendering.V2.Cache;
+using Beutl.Graphics.Rendering.Cache;
 using Beutl.Logging;
 using Beutl.Media;
 using Beutl.Media.Music;
