+++ tests/Beutl.UnitTests/Engine/GraphicsContext2DTests.cs
@@ -1,6 +1,6 @@
 ﻿using Beutl.Graphics;
 using Beutl.Graphics.Effects;
-using Beutl.Graphics.Rendering.V2;
+using Beutl.Graphics.Rendering;
 using Beutl.Graphics.Shapes;
 using Beutl.Graphics.Transformation;
 using Beutl.Logging;
