+++ tests/Beutl.UnitTests/ProjectSystem/Operation/PublishOperatorTests.cs
@@ -1,6 +1,6 @@
 ﻿using Beutl.Graphics;
 using Beutl.Graphics.Effects;
-using Beutl.Graphics.Rendering.V2;
+using Beutl.Graphics.Rendering;
 using Beutl.Graphics.Transformation;
 using Beutl.Media;
 using Beutl.Operation;
